+++ common/src/main/java/net/wizards/client/effect/FrostShieldRenderer.java
@@ -16,7 +16,7 @@ public class FrostShieldRenderer implements CustomModelStatusEffect.Renderer {
     public static final Identifier modelId_base = new Identifier(WizardsMod.ID, "effect/frost_shield_base");
     public static final Identifier modelId_overlay = new Identifier(WizardsMod.ID, "effect/frost_shield_overlay");
 
-    private static final RenderLayer BASE_RENDER_LAYER = RenderLayer.getTranslucentMovingBlock();
+    private static final RenderLayer BASE_RENDER_LAYER = RenderLayer.getEntityTranslucent(SpriteAtlasTexture.BLOCK_ATLAS_TEXTURE);
     private static final RenderLayer OVERLAY_RENDER_LAYER = CustomLayers.projectile(SpriteAtlasTexture.BLOCK_ATLAS_TEXTURE, false);
 
     @Override