+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.9.14
 
 - Reduce cost of netherite staves based on community feedback
+- Phase out deprecated API usage
 
 # 0.9.13